Add multi-stage creation wait
This extends the creator with waitStages for resources which want to
wait on any other dependent resources in the whole creation flow. This
also reworks the printing system as the previous approach was a bit
flawed. The spinner is now on the line that is currently in progress and
is then updated to a tick/cross depending on the outcome of the stage.