Get proper token for log client
Because our Kubeconfig uses an exec-based auth provider, we need to get
the token in a different way. This moves the relevant code to execute
the auth plugin programmatically to the api package to avoid import
cycles. While at it this tries to improve it a bit and not construct
the whole OIDCCmd and execute it through Kong, which always felt a
bit wrong.