do not set defaults in the application on creation
If we set defaults in the application on creation, existing content in the .deploio.yaml file will be ignored (as the config which is set on the Application itself has higher precedence). As we will always fall back to valid defaults, we can just leave them out in the application if the user did not set an explicit value.