feat: add dynamic completion
This adds initial support for dynamic completion. For example `nctl update
application <tab><tab>` will return a list of applications in the
current project. There are some edge-cases where the arg/flag name does
not match the resource name. To make these cases work, they can be added
to a map in the predictor.