fix: improve logbox interrupt handling
This fixes a number of issues with the cancel/error handling while the
logbox (tea program) is running. Previously input was sent to the tea
program, which meant we had to handle ctrl+c/ctrl+d explicitly. Now the
tea program is stopped using the context and also in afterWait since
it's important to call `p.Wait()` so that the console is restored
properly before nctl exits.