fix: use dynamic auth by default
In a previous PR, we changed the client to initialize the token once
during creation of the client. This is problematic for longer-running
processes such as creating an app where the token might expire at
runtime. The only reason for putting the token into the client config
was an issue with the autocompletion where the dynamic exec config
fails for some shells.