feat: add labels to logs output
This adds a select set of labels to the app/build output. Also fixes a
bug in the tailQuery function where labels would be appended instead of
created for ever log line received.

The app logs now show all logs related to an app by default. The output
can be filtered by type if desired.