feat: warn when setting project/org
When setting the project or organization, it will now warn you if the
project does not exist, your current user does not have access to it or
if the organization is not in the list of available organizations. We
don't error out to still support some use-cases where a user might not
have the permission to get a project or might not be fully part of an
org but might still have access to some resource in it.