fix: improve error log on app creation
When a build fails and we fetch the last x lines of the build log and
just wait until *something* is returned by the logs service. But fresh
(and relevant) logs might still be streaming in after the build has
already been declared as failed, so we might be missing the actual build
error. Instead, we now use the tail function with a short-lived context.
We are tailing for 5 seconds, which should be enough to get the last
lines of the build/release log. Another benefit is that the tailing
function already has retries built-in so we don't need to wrap it in a
retry like before.