feat: add watch flag to get commands
This adds a new `--watch` flag that works on most get commands. It does
a normal get and then keeps nctl running while watching the specified
resource for changes. This also works in combination with output options
such as yaml, json etc.

As long as the resource is using the `get.list()` API, watch will
automatically work. To accomplish this, the get commands need to
implement a common interface `listPrinter`. For now, watch does not work
in combination with the --all-projects and --all-namespaces flags.